Add new tokens needed for JSX support
As a Python easter egg, '<>' is an alias for '!='. Since we need '<>'
for JSX fragments, we make '<>' a separate token. This breaks backwards
compatibility with Barry's syntax.. Literally unsuable.

We also add a new token for identifier. There already is a token for it
(called NAME) but it directly converts the token into `ast.Name` which
we don't want when dealing with JSX component names. For use-provided
components, it should become `ast.Name` but for HTML built-ins like
`<div />` the identifier should become a string constant instead. This
is easier to do when operating on the raw NAME token (Token*) which is
what the NAME_RAW token returns.