Add grammar for JSX syntax
What is supported:

- JSX elements:

    <div></div>

- Self-closing JSX elements:

    <div />

- JSX Fragments:

    <>
        <div></div>
        <div></div>
    </>

- Props:

    <div id="foo" className="bar"></div>

- Keyword-only props:

    <div disabled></div>

- Expressions:

    <div>{foo}</div>
    <div prop={1+1} />

What is not supported:

- Text nodes must be wrapped in quotes (i.e. be valid Python strings):

    <div>"foo"</div>

This is illegal:

    <div>foo</div>

This is a limitation of the way the parser and tokenizer work together.
It is in theory possible to make this work, but it'd require some
additional work on low-level C code which I'd rather avoid ;)

- Spread attributes:

    <div {...props}></div>

This probably could be supported by just playing with the grammar a bit.